Solve clash of TFT and buzzer libraries
The TFT's reset pin is not connected to any GPIO.  However, the TFT library requires a RST pin to be set.  The current TFT library had IO32 set as reset pin.  However, this pin is used for the buzzer.  Hence, the buzzer stopped working if the TFT library is loaded as well.

This is solved by setting the reset pin to IO26 which is not connected to anything on the board (though it is still broken out on the microbit connector).