Add methods to enable and disable the TFT backlight
Note that this will only work if the backlight switch is set to AUTO.
Note also that the select button on the game-on will not work if the
backlight is disabled.

The way this is implemented is by using a trick.  The backlight LED is
not directly connected to any pin of the ESP32, but it is connected to
the interrupt pins of the accelerometer.  The level of IO36 can
therefore be manipulated by enabling/disabling the interrupt
functionality on the accelerometer, which will pull the line HIGH or
LOW.